re-work folder's structure, improved Context
