--- src/patterns/strategy/context.impl.ts
@@ -1,13 +1,0 @@
-import {IBaseDiscountStrategy, IStrategyContext} from "./interfaces";
-
-export class StrategyContext implements IStrategyContext{
-  private strategy: IBaseDiscountStrategy;
-
-  constructor(strategy: IBaseDiscountStrategy) {
-    this.strategy = strategy;
-  }
-
-  calculateFinalPrice(price: number): number {
-    return this.strategy.calculateDiscount(price);
-  }
-}
