+++ src/patterns/strategy/discount-example/context/interfaces.ts
@@ -1,0 +1,8 @@
+export interface IDiscountStrategy {
+  name: string;
+  calculateDiscount(price: number): number;
+}
+
+export interface IStrategyContext {
+  calculateFinalPrice(price: number): number;
+}
