+++ src/patterns/strategy/discount-example/fixed-discount/implementation.ts
@@ -1,0 +1,11 @@
+import {IFixedDiscountStrategy} from "./interfaces";
+
+export class FixedDiscount implements IFixedDiscountStrategy {
+  public name = 'FixedDiscount';
+
+  constructor(public amount: number) {}
+
+  calculateDiscount(price: number): number {
+    return price - this.amount;
+  }
+}
