+++ src/patterns/strategy/discount-example/fixed-discount/interfaces.ts
@@ -1,0 +1,5 @@
+import {IDiscountStrategy} from "../context/interfaces";
+
+export interface IFixedDiscountStrategy extends IDiscountStrategy {
+  amount: number;
+}
