+++ src/patterns/strategy/discount-example/percentage-discount/implementation.ts
@@ -1,0 +1,12 @@
+import {IPercentageDiscountStrategy} from "./interfaces";
+
+export class PercentageDiscount implements IPercentageDiscountStrategy {
+  public name = 'PercentageDiscount';
+
+  constructor(public percentage: number) {
+  }
+
+  calculateDiscount(price: number): number {
+    return price - (price * this.percentage) / 100;
+  }
+}
