+++ src/patterns/strategy/discount-example/percentage-discount/interfaces.ts
@@ -1,0 +1,5 @@
+import {IDiscountStrategy} from "../context/interfaces";
+
+export interface IPercentageDiscountStrategy extends IDiscountStrategy {
+  percentage: number;
+}
