--- src/patterns/strategy/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces'
-export * from './strategies.impl'
-export * from './context.impl'
