+++ src/patterns/strategy/strategy.md
@@ -1,4 +1,6 @@
-# "Strategy" Pattern Explanation:
-The Strategy pattern allows you to define a family of algorithms, encapsulate each one, and make them replaceable.
-It lets the algorithm vary independently of the clients that use it.
-A typical use case would be implementing different **sorting** or **pricing** strategies.
+# Strategy Pattern Explanation:
+The Strategy pattern allows you to define a family of algorithms, </br>
+encapsulate each one, and make them replaceable. </br>
+It lets the algorithm vary independently of the clients that use it.</br>
+
+A typical use case would be implementing different **sorting** or **pricing** strategies.</br>
